+++ CHANGELOG.md
@@ -25,6 +25,8 @@
   - [#4617](https://github.com/wix/ricos/pull/4617) fix `isAdsense` deprecation in `htmlData`
 - `link`
   - [#4616](https://github.com/wix/ricos/pull/4616) add support to `ANCHOR` type to avoid normaliztion in viewer
+- `ricos-editor`
+  - [#4629](https://github.com/wix/ricos/pull/4629) move `rceNext` to `RicosEditorProps`
 
 </details>